move Application to its own module
